Added getDeployments method, moved around reflection code for Greengrass APIs
